Move unpack_py_instruction to py_methods
